fix: update timestamp type from uint64 to uint256 across codebase (#21)
* fix: update timestamp type from uint64 to uint256 across codebase

Match the registered EAS schemas on Base Sepolia which use uint256
for timestamps. This ensures schema registration scripts, tests,
documentation, and example contracts are consistent with the actual
deployed schemas.

Files updated:
- scripts/register-schemas.ts and .mjs
- integration test decode types
- SPEC.md schema definitions
- example contracts and todos

* fix: add missing build scripts to workspace packages

+++ examples/contracts/LocationGatedNFT.sol
@@ -68,10 +68,10 @@ contract LocationGatedNFT is SchemaResolver {
         }
 
         // Decode the boolean policy attestation
-        // Schema: "bool result, bytes32[] inputRefs, uint64 timestamp, string operation"
+        // Schema: "bool result, bytes32[] inputRefs, uint256 timestamp, string operation"
         (bool result, , , ) = abi.decode(
             attestation.data,
-            (bool, bytes32[], uint64, string)
+            (bool, bytes32[], uint256, string)
         );
 
         // Require the location check to have passed
